+++ android/app/src/test/java/org/tncsim/twa/PackageIdentityTest.java
@@ -1,0 +1,13 @@
+package org.tncsim.twa;
+
+import static org.junit.Assert.assertEquals;
+
+import org.junit.Test;
+
+public class PackageIdentityTest {
+
+    @Test
+    public void applicationIdKeepsExistingPlayListing() {
+        assertEquals("org.tncsim.twa", MainActivity.class.getPackage().getName());
+    }
+}